More review comment changes
